New/dice-app/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
python-dotenv==1.0.1